@@ -1,12 +1,11 @@
 ﻿using AutoMapper;
 
-namespace WebAPP.Infrastructure.Infrastructure
-{
-    public class AccountProfile : Profile
-    {
-        public AccountProfile()
-        {
+namespace WebAPP.Infrastructure;
 
-        }
+public class AccountProfile : Profile
+{
+    public AccountProfile()
+    {
+
     }
 }
