@@ -1,16 +1,14 @@
-﻿using AutoMapper;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.Extensions.Localization;
 using System.Net;
-using WebAPP.Extensions;
-using WebAPP.Infrastructure.GlobalParameters;
-using WebAPP.Infrastructure.Models.enums;
-using WebAPP.MiddlewareFactory;
-using WebAPP.Infrastructure.Utilities;
 using System.Text.Json;
 using WebApp.Infrastructure.Models.dto;
 using WebApp.Infrastructure.Utilities;
+using WebAPP.Extensions;
+using WebAPP.Infrastructure;
+using WebAPP.Infrastructure.Models.enums;
+using WebAPP.MiddlewareFactory;
 using WebAPP.Services;
 
 namespace WebAPP.Controllers
