@@ -8,8 +8,8 @@ using NLog.Web;
 using System.Globalization;
 using WebAPP.Controllers;
 using WebAPP.Extensions;
+using WebAPP.Infrastructure;
 using WebAPP.Infrastructure.DBContext;
-using WebAPP.Infrastructure.GlobalParameters;
 using WebAPP.Infrastructure.Models.enums;
 using WebAPP.Infrastructure.Models.Validation;
 using WebAPP.Infrastructure.Utilities;
