@@ -25,7 +25,7 @@
 <p class="mt-4 mb-2 text-body-secondary"></p>
 
 <script type="text/javascript">
-    $(document).ready(function () {
+    $(function () {
         let params = new URLSearchParams(document.location.search);
         @{
             string message = string.Empty;
